+++ cli/oci_test.go
@@ -317,106 +317,3 @@ func TestGetCgroupsDirPath(t *testing.T) {
 		assert.Equal(d.expectedResult, path)
 	}
 }
-
-func TestFetchContainerIDMappingContainerIDEmptyFailure(t *testing.T) {
-	assert := assert.New(t)
-
-	sandboxID, err := fetchContainerIDMapping("")
-	assert.Error(err)
-	assert.Empty(sandboxID)
-}
-
-func TestFetchContainerIDMappingEmptyMappingSuccess(t *testing.T) {
-	assert := assert.New(t)
-
-	path, err := ioutil.TempDir("", "containers-mapping")
-	assert.NoError(err)
-	defer os.RemoveAll(path)
-	ctrsMapTreePath = path
-
-	sandboxID, err := fetchContainerIDMapping(testContainerID)
-	assert.NoError(err)
-	assert.Empty(sandboxID)
-}
-
-func TestFetchContainerIDMappingTooManyFilesFailure(t *testing.T) {
-	assert := assert.New(t)
-
-	path, err := createTempContainerIDMapping(testContainerID, testSandboxID)
-	assert.NoError(err)
-	defer os.RemoveAll(path)
-	err = os.MkdirAll(filepath.Join(ctrsMapTreePath, testContainerID, testSandboxID+"2"), ctrsMappingDirMode)
-	assert.NoError(err)
-
-	sandboxID, err := fetchContainerIDMapping(testContainerID)
-	assert.Error(err)
-	assert.Empty(sandboxID)
-}
-
-func TestFetchContainerIDMappingSuccess(t *testing.T) {
-	assert := assert.New(t)
-
-	path, err := createTempContainerIDMapping(testContainerID, testSandboxID)
-	assert.NoError(err)
-	defer os.RemoveAll(path)
-
-	sandboxID, err := fetchContainerIDMapping(testContainerID)
-	assert.NoError(err)
-	assert.Equal(sandboxID, testSandboxID)
-}
-
-func TestAddContainerIDMappingContainerIDEmptyFailure(t *testing.T) {
-	assert := assert.New(t)
-
-	err := addContainerIDMapping(context.Background(), "", testSandboxID)
-	assert.Error(err)
-}
-
-func TestAddContainerIDMappingSandboxIDEmptyFailure(t *testing.T) {
-	assert := assert.New(t)
-
-	err := addContainerIDMapping(context.Background(), testContainerID, "")
-	assert.Error(err)
-}
-
-func TestAddContainerIDMappingSuccess(t *testing.T) {
-	assert := assert.New(t)
-
-	path, err := ioutil.TempDir("", "containers-mapping")
-	assert.NoError(err)
-	defer os.RemoveAll(path)
-	ctrsMapTreePath = path
-
-	_, err = os.Stat(filepath.Join(ctrsMapTreePath, testContainerID, testSandboxID))
-	assert.True(os.IsNotExist(err))
-
-	err = addContainerIDMapping(context.Background(), testContainerID, testSandboxID)
-	assert.NoError(err)
-
-	_, err = os.Stat(filepath.Join(ctrsMapTreePath, testContainerID, testSandboxID))
-	assert.NoError(err)
-}
-
-func TestDelContainerIDMappingContainerIDEmptyFailure(t *testing.T) {
-	assert := assert.New(t)
-
-	err := delContainerIDMapping(context.Background(), "")
-	assert.Error(err)
-}
-
-func TestDelContainerIDMappingSuccess(t *testing.T) {
-	assert := assert.New(t)
-
-	path, err := createTempContainerIDMapping(testContainerID, testSandboxID)
-	assert.NoError(err)
-	defer os.RemoveAll(path)
-
-	_, err = os.Stat(filepath.Join(ctrsMapTreePath, testContainerID, testSandboxID))
-	assert.NoError(err)
-
-	err = delContainerIDMapping(context.Background(), testContainerID)
-	assert.NoError(err)
-
-	_, err = os.Stat(filepath.Join(ctrsMapTreePath, testContainerID, testSandboxID))
-	assert.True(os.IsNotExist(err))
-}
